feat: add items structured data (#44)
* add structured data

* add raw item fields

* add words to dictionary

* add fields to item

* add recycling data endpoint

* update discriminator values

* add new raw data types

* add services

* add settings

* fix build

* fix warning

* fix write

* post merge

* fix deserialization of doubles

* allow writing NaNs

* add jobs endpoints

* reduce size of recipe dto

* put get recipes endpoint back

+++ DataCenter/Raw/Services/Items/RawItemSetsService.cs
@@ -8,6 +8,6 @@ public class RawItemSetsService(IReadOnlyCollection<RawItemSet> itemSets)
 {
     readonly Dictionary<int, RawItemSet> _itemSets = itemSets.ToDictionary(itemSet => itemSet.Id, itemSet => itemSet);
 
-    public RawItemSet? GetItemSet(ushort itemSetId) => _itemSets.GetValueOrDefault(itemSetId);
+    public RawItemSet? GetItemSet(int itemSetId) => _itemSets.GetValueOrDefault(itemSetId);
     public IEnumerable<RawItemSet> GetItemSets() => _itemSets.Values;
 }
